+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # NEXT RELEASE
+* Put in a cache busting fallback if user's RecentUserToots end up empty
 
 ### v0.77.4
-* Go back to manual Toot construction in buildFromApiObjects()
+* Go back to manual `Toot` construction in `MastoApi.buildFromApiObjects()`
 
 ### v0.77.3
 * Logging of sortByValue results